gpg/requirements.txt
kubiya
pydantic
requests
python-gnupg
PyGithub